requirements.txt
absl-py==0.2.2
astor==0.6.2
astroid==1.6.4
bleach==1.5.0
dm-sonnet==1.20
gast==0.2.0
grpcio==1.12.0
html5lib==0.9999999
isort==4.3.4
lazy-object-proxy==1.3.1
Markdown==2.6.11
mccabe==0.6.1
numpy==1.14.3
protobuf==3.5.2.post1
pylint==1.9.1
six==1.11.0
tensorboard==1.8.0
tensorflow==1.8.0
termcolor==1.1.0
tqdm==4.23.4
Werkzeug==0.14.1
wrapt==1.10.11
yapf==0.22.0
